split alfa_node src file into multiple files to improve readability
